fix(forge): prevent 0 range loading

+++ forge/src/main/java/com/envyful/holograms/forge/hologram/database/JsonHologramSaver.java
@@ -90,7 +90,7 @@ public class JsonHologramSaver implements HologramSaver {
         int range = 64;
 
         if (map.containsKey("range")) {
-            range = (int) map.get("range");
+            range = (int) ((double) map.get("range"));
 
             if (range <= 0) {
                 range = 64;
